refactor: renaming configs/sidebarOptions to configs/sidebar
